Initial work on loading map from file

+++ lidar.cpp
@@ -85,7 +85,7 @@ std::vector<LidarRay> lidar::GetLidarPoints(ConstLaserScanStampedPtr &msg)
     lidarRays.clear();
 
     // Loop through each scan ray
-    for (size_t i = 0; i < msg->scan().ranges_size(); i++)
+    for (int i = 0; i < msg->scan().ranges_size(); i++)
     {
         // Get angle and distance of the ray
         float angle = angleMin + i * angleIncrement;
